[feature] complete replacement of the theme, including color, shape and type into material 3
