+++ library/designsystem/src/main/java/ir/composenews/designsystem/theme/Color.kt
@@ -1,12 +1,71 @@
 package ir.composenews.designsystem.theme
-
 import androidx.compose.ui.graphics.Color
 
-val Red200 = Color(0xfff297a2)
-val Red300 = Color(0xffea6d7e)
-val Red700 = Color(0xffdd0d3c)
-val Red800 = Color(0xffd00036)
-val Red900 = Color(0xffc20029)
+val md_theme_light_primary = Color(0xffdd0d3c)
+val md_theme_light_onPrimary = Color(0xFFFFFFFF)
+val md_theme_light_primaryContainer = Color(0xFFFFDADC)
+val md_theme_light_onPrimaryContainer = Color(0xFF400010)
+val md_theme_light_secondary = Color(0xffdd0d3c)
+val md_theme_light_onSecondary = Color(0xFFFFFFFF)
+val md_theme_light_secondaryContainer = Color(0xFFFFDAD9)
+val md_theme_light_onSecondaryContainer = Color(0xFF40000A)
+val md_theme_light_tertiary = Color(0xFF984060)
+val md_theme_light_onTertiary = Color(0xFFFFFFFF)
+val md_theme_light_tertiaryContainer = Color(0xFFFFD9E2)
+val md_theme_light_onTertiaryContainer = Color(0xFF3E001D)
+val md_theme_light_error = Color(0xFFBA1A1A)
+val md_theme_light_errorContainer = Color(0xFFFFDAD6)
+val md_theme_light_onError = Color(0xFFFFFFFF)
+val md_theme_light_onErrorContainer = Color(0xFF410002)
+val md_theme_light_background = Color(0xFFFFFBFF)
+val md_theme_light_onBackground = Color(0xFF201A1A)
+val md_theme_light_surface = Color(0xFFFFFBFF)
+val md_theme_light_onSurface = Color(0xFF201A1A)
+val md_theme_light_surfaceVariant = Color(0xFFF4DDDE)
+val md_theme_light_onSurfaceVariant = Color(0xFF524344)
+val md_theme_light_outline = Color(0xFF857374)
+val md_theme_light_inverseOnSurface = Color(0xFFFBEEEE)
+val md_theme_light_inverseSurface = Color(0xFF362F2F)
+val md_theme_light_inversePrimary = Color(0xFFFFB2B9)
+val md_theme_light_shadow = Color(0xFF000000)
+val md_theme_light_surfaceTint = Color(0xFFA4384A)
+val md_theme_light_outlineVariant = Color(0xFFD7C1C2)
+val md_theme_light_scrim = Color(0xFF000000)
 
+val md_theme_dark_primary = Color(0xffea6d7e)
+val md_theme_dark_onPrimary = Color(0xFF000000)
+val md_theme_dark_primaryContainer = Color(0xFF852034)
+val md_theme_dark_onPrimaryContainer = Color(0xFFFFDADC)
+val md_theme_dark_secondary = Color(0xffea6d7e)
+val md_theme_dark_onSecondary = Color(0xFF000000)
+val md_theme_dark_secondaryContainer = Color(0xFF920023)
+val md_theme_dark_onSecondaryContainer = Color(0xFFFFDAD9)
+val md_theme_dark_tertiary = Color(0xFFFFB1C7)
+val md_theme_dark_onTertiary = Color(0xFF5E1132)
+val md_theme_dark_tertiaryContainer = Color(0xFF7B2948)
+val md_theme_dark_onTertiaryContainer = Color(0xFFFFD9E2)
+val md_theme_dark_error = Color(0xFFFFB4AB)
+val md_theme_dark_errorContainer = Color(0xFF93000A)
+val md_theme_dark_onError = Color(0xFF690005)
+val md_theme_dark_onErrorContainer = Color(0xFFFFDAD6)
+val md_theme_dark_background = Color(0xFF201A1A)
+val md_theme_dark_onBackground = Color(0xFFECE0E0)
+val md_theme_dark_surface = Color(0xFF201A1A)
+val md_theme_dark_onSurface = Color(0xFFECE0E0)
+val md_theme_dark_surfaceVariant = Color(0xFF524344)
+val md_theme_dark_onSurfaceVariant = Color(0xFFD7C1C2)
+val md_theme_dark_outline = Color(0xFF9F8C8D)
+val md_theme_dark_inverseOnSurface = Color(0xFF201A1A)
+val md_theme_dark_inverseSurface = Color(0xFFECE0E0)
+val md_theme_dark_inversePrimary = Color(0xFFA4384A)
+val md_theme_dark_shadow = Color(0xFF000000)
+val md_theme_dark_surfaceTint = Color(0xFFFFB2B9)
+val md_theme_dark_outlineVariant = Color(0xFF524344)
+val md_theme_dark_scrim = Color(0xFF000000)
+
+
+val seed = Color(0xFFEA6D7E)
+
+// custom colors
 val graphColor = Color(0xFF6750A4)
 val lightGraphColor = graphColor.copy(alpha = 0.5f)
